This is a working version of a face recognition app using Clarifai API
This change will commit the frontend-side code of a full stack web application that recognizes faces, with multiple features such as user authentication and user score count of how many faces they have detected so far. This app is developed using React framework and using Clarifai's face detection API. This is a significant commit since the past changes has not been consistantly committed. There will be barely any commit history. App.js handles the calls to necessary components, and the components are kept in the components directory. There are three separate pages that a user could navigate to: sign-in, register, and the main page. In the main page, a user could enter a link to a image in jpg format and this app will detect and surround a face with a blue box to indicate the detection of a face. An user could register a new account using their name, email, and password. Once registered, the same user could use their email and password to signin. The app keeps track of each user's face-detection-count, and displays it in the main page. As of now, the app only detect one face even if there are multiple faces in an input picture.