Added functionality to increment face count based on the number of faced detected.
Previously, face count only incremented by 1 regardless of the number of faced detected. This is now fixed by counting the number of entries that was returned from the API call and storing each information about the location of face into an state array. This array of of face location is then passed to FaceRecognition.js and displayed as blue boxes. A couple of errors are fixed, including fixing  an error showing when the detect button was pressed when there is no input and an error when there was no data returned after an API call (happens when there is no faces detected in input picture).